Fixed demo bugs, dates are now formatted, removed @Jsonignore on id in dtos, get all method, date of modification only in update operation ,no tag property = empty tags, 204 no content in delete

+++ Task_1/service/src/main/java/com/epam/esm/service/implementation/GiftCertificateServiceImpl.java
@@ -11,6 +11,7 @@ import com.epam.esm.exception.ResourceNotFoundException;
 import com.epam.esm.repository.BaseCRUDRepository;
 import com.epam.esm.repository.GiftCertificateRepository;
 import com.epam.esm.repository.TagRepository;
+import com.epam.esm.repository.specfication.FindAllCertificates;
 import com.epam.esm.repository.specfication.FindGiftCertificateByID;
 import com.epam.esm.repository.specfication.FindGiftCertificatesByDescription;
 import com.epam.esm.repository.specfication.FindGiftCertificatesByName;
@@ -27,6 +28,7 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -54,7 +56,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
             GiftCertificate giftCertificate = giftCertificateRepo.findByID(id);
             GiftCertificateDTO giftCertificateDTO = giftCertificateConverter.toDTO(
                     giftCertificate);
-            giftCertificateDTO.setTagDTOList(
+            giftCertificateDTO.setTags(
                     tagConverter.toDTOList(tagRepository.query(
                             new FindTagsByCertificateID(giftCertificateDTO.getId()))));
             return giftCertificateDTO;
@@ -62,6 +64,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
             throw new ResourceNotFoundException("Gift Certificate with ID: " + id + " was not found!");
         }
     }
+
 
     public List<GiftCertificateDTO> getGiftCertificate(Long tagID, String name, String desc, Integer sortByDate, Integer sortByName) {
         List<Specification<GiftCertificate>> specifications = new ArrayList<>();
@@ -100,10 +103,24 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
         for (GiftCertificateDTO i : dtoResult) {
             List<TagDTO> tags = tagConverter.toDTOList(tagRepository.query(
                     new FindTagsByCertificateID(i.getId())));
-            i.setTagDTOList(tags);
+            i.setTags(tags);
             dtoResultWithTags.add(i);
         }
         return dtoResultWithTags;
+    }
+
+    @Override
+    public List<GiftCertificateDTO> getAllGiftCertificates() {
+        try {
+            List<GiftCertificateDTO> dtos = giftCertificateConverter.toDTOList(
+                    giftCertificateRepo.query(new FindAllCertificates()));
+            for (GiftCertificateDTO i : dtos) {
+                i.setTags(tagConverter.toDTOList(tagRepository.query(new FindTagsByCertificateID(i.getId()))));
+            }
+            return dtos;
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException("Didn't find gift certificates");
+        }
     }
 
     @Transactional
@@ -111,30 +128,42 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     public GiftCertificateDTO add(GiftCertificateDTO certificateDTO) {
         addNewTagsIfTheyDontExist(certificateDTO);
         GiftCertificate certificate = giftCertificateConverter.toEntity(certificateDTO);
-        GiftCertificate giftCertificateCreated = giftCertificateRepo.add(certificate);
-        certificateDTO.setId(giftCertificateCreated.getId());
+        certificate.setDateOfCreation(LocalDate.now());
+        certificate.setDateOfModification(null);
+        certificate = giftCertificateRepo.add(certificate);
+        certificateDTO.setId(certificate.getId());
         addReferences(certificateDTO);
-        GiftCertificateDTO giftCertificateDTOCreated = giftCertificateConverter.toDTO(giftCertificateCreated);
+        GiftCertificateDTO giftCertificateDTOCreated = giftCertificateConverter.toDTO(certificate);
         List<TagDTO> tagDTOList = tagConverter.toDTOList(tagRepository.query(
                 new FindTagsByCertificateID(giftCertificateDTOCreated.getId())));
-        giftCertificateDTOCreated.setTagDTOList(tagDTOList);
+        if (tagDTOList != null) {
+            giftCertificateDTOCreated.setTags(tagDTOList);
+        }
         return giftCertificateDTOCreated;
     }
 
     @Transactional
     @Override
     public GiftCertificateDTO update(GiftCertificateDTO certificateDTO) {
-        List<TagDTO> newTagsDTO = addNewTagsIfTheyDontExist(certificateDTO);
+        if (getByID(certificateDTO.getId()) == null) {
+            throw new ResourceNotFoundException("Gift Certificate with ID: "
+                    + certificateDTO.getId() + " was not found!");
+        }
+        addNewTagsIfTheyDontExist(certificateDTO);
         GiftCertificate certificate = giftCertificateConverter.toEntity(certificateDTO);
+        certificate.setDateOfModification(LocalDate.now());
         giftCertificateRepo.update(certificate);
-        certificateDTO.setId(certificate.getId());
-        setNewTagsIDinDTO(certificateDTO, newTagsDTO);
+        List<TagDTO> newTags = certificateDTO.getTags();
+        for (TagDTO i : newTags) {
+            i.setId(tagRepository.findByName(i.getName()).getId());
+        }
+        certificateDTO.setTags(newTags);
         updateReferences(certificateDTO);
         return getByID(certificateDTO.getId());
     }
 
-    private void setNewTagsIDinDTO(GiftCertificateDTO certificateDTO, List<TagDTO> newTagsDTO) {
-        List<TagDTO> tagDTOList = certificateDTO.getTagDTOList();
+   /* private void setNewTagsIDinDTO(GiftCertificateDTO certificateDTO, List<TagDTO> newTagsDTO) {
+        List<TagDTO> tagDTOList = certificateDTO.getTags();
         List<TagDTO> tagDTOListWithoutNullElements = new ArrayList<>();
         for (TagDTO i : tagDTOList) {
             if (i.getId() == null) {
@@ -149,12 +178,16 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
                 tagDTOListWithoutNullElements.add(i);
             }
         }
-        certificateDTO.setTagDTOList(tagDTOListWithoutNullElements);
-    }
+        certificateDTO.setTags(tagDTOListWithoutNullElements);
+    }*/
 
     @Transactional
     @Override
     public boolean delete(long id) {
+        if (getByID(id) == null) {
+            throw new ResourceNotFoundException("Gift Certificate with ID: "
+                    + id + " was not found!");
+        }
         giftCertificateRepo.delete(new FindGiftCertificateByID(id));
         return true;
     }
@@ -162,47 +195,58 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     @Transactional
     @Override
     public boolean delete(GiftCertificateDTO certificate) {
+        if (getByID(certificate.getId()) == null) {
+            throw new ResourceNotFoundException("Gift Certificate with ID: "
+                    + certificate.getId() + " was not found!");
+        }
         giftCertificateRepo.delete(giftCertificateConverter.toEntity(certificate));
         return true;
     }
 
     private List<TagDTO> addNewTagsIfTheyDontExist(GiftCertificateDTO dto) {
-        List<TagDTO> tags = dto.getTagDTOList();
-        List<Tag> newTags = new ArrayList<>();
         List<Tag> newTagsWithID = new ArrayList<>();
-        for (TagDTO i : tags) {
-            List<Tag> tagFound = tagRepository.query(new FindTagByName(i.getName()));
-            if (tagFound.isEmpty()) {
-                newTags.add(tagConverter.toEntity(i));
+        if (dto.getTags() != null) {
+            List<TagDTO> tags = dto.getTags();
+            List<Tag> newTags = new ArrayList<>();
+            for (TagDTO i : tags) {
+                List<Tag> tagFound = tagRepository.query(new FindTagByName(i.getName()));
+                if (tagFound.isEmpty()) {
+                    newTags.add(tagConverter.toEntity(i));
+                }
             }
-        }
-        for (Tag i : newTags) {
-            newTagsWithID.add(tagRepository.add(i));
+            for (Tag i : newTags) {
+                newTagsWithID.add(tagRepository.add(i));
+            }
         }
         return tagConverter.toDTOList(newTagsWithID);
     }
 
     private void addReferences(GiftCertificateDTO dto) {
-        GiftCertificate certificate = giftCertificateConverter.toEntity(dto);
-        List<TagDTO> tagsDTO = dto.getTagDTOList();
-        List<Tag> tags = tagConverter.toEntityList(tagsDTO);
-        GiftCertificateRepository repo = (GiftCertificateRepository) giftCertificateRepo;
-        for (Tag i : tags) {
-            repo.addReference(certificate, i);
+        if (dto.getTags() != null) {
+            GiftCertificate certificate = giftCertificateConverter.toEntity(dto);
+            List<TagDTO> tagsDTO = dto.getTags();
+            List<Tag> tags = tagConverter.toEntityList(tagsDTO);
+            GiftCertificateRepository repo = (GiftCertificateRepository) giftCertificateRepo;
+            for (Tag i : tags) {
+                i.setId(tagRepository.findByName(i.getName()).getId());
+                repo.addReference(certificate, i);
+            }
         }
+
     }
 
     private void updateReferences(GiftCertificateDTO dto) {
-        GiftCertificate certificate = giftCertificateConverter.toEntity(dto);
-        List<TagDTO> newTagsDTO = dto.getTagDTOList();
-        List<Tag> newTags = tagConverter.toEntityList(newTagsDTO);
-        List<Tag> oldTags = tagRepository.query(new FindTagsByCertificateID(certificate.getId()));
-        newTags.removeAll(oldTags);
-        GiftCertificateRepository repo = (GiftCertificateRepository) giftCertificateRepo;
-        for (Tag i : newTags) {
-            repo.addReference(certificate, i);
+        if (dto.getTags() != null) {
+            GiftCertificate certificate = giftCertificateConverter.toEntity(dto);
+            List<TagDTO> newTagsDTO = dto.getTags();
+            List<Tag> newTags = tagConverter.toEntityList(newTagsDTO);
+            GiftCertificateRepository repo = (GiftCertificateRepository) giftCertificateRepo;
+            repo.deleteAllReferences(certificate);
+            for (Tag i : newTags) {
+                i.setId(tagRepository.findByName(i.getName()).getId());
+                repo.addReference(certificate, i);
+            }
         }
     }
-
 
 }
